requirements.txt
fastapi
streamlit
google-generativeai
PyPDF2
python-dotenv
pylance
dotenv
